requirements.txt

fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
google-generativeai==0.8.3
ibm-watsonx-ai==1.1.21
requests==2.31.0
aiohttp==3.9.3
python-logging-loki==0.3.1
tenacity==8.2.3
